// src/transactions/enums/transaction.enum.ts
export enum TransactionType {
    SALES = 'SALES',
    EXPENSES = 'EXPENSES'
}

export enum TransactionStatus {
    COMPLETED = 'COMPLETED',
    PENDING = 'PENDING'
}

export enum Region {
    NORTH_AMERICA = 'NORTH_AMERICA',
    SOUTH_AMERICA = 'SOUTH_AMERICA',
    EUROPE = 'EUROPE',
    ASIA = 'ASIA',
    AFRICA = 'AFRICA',
    OCEANIA = 'OCEANIA'
}

export enum Category {
    HR = 'HR',
    MARKETING = 'MARKETING',
    OPERATIONS = 'OPERATIONS'
} 